refactor: adapted player and order class to be entities

+++ src/main/java/fr/univ_smb/isc/m1/trading_game/infrastructure/persistence/Game.java
@@ -7,7 +7,7 @@ import java.util.*;
 public class Game {
     @Id
     @GeneratedValue
-    protected int id;
+    protected long id;
     protected int maxPortfolios;
     protected int initialBalance; //cents
     protected int transactionFee; //cents
